Added new sorting to Nursing:
<topic name="none">
  <pattern match="HPRS 1203" />
  <pattern match="RNSG 1211" />
  <pattern match="RNSG 1301" />
  <pattern match="RNSG 1108" />
</topic>
<topic name="First Semester Program Students">
  <pattern match="RNSG 1105" />
  <pattern match="RNSG 1413" />
  <pattern match="RNSG 1360" />
</topic>
<topic name="Third Semester Program Students">
  <pattern match="RNSG 1443" />
  <pattern match="RNSG 2461" />
</topic>

This also necessitated changing the base.xml file's nursing subject to <subject name="Nursing - Unsorted"> and moving the comments over to the spring.xml file.

Of course, this left Fall with only the unsorted nursing subject. So, fall has a subject nursing identical to the old base.xml nursing subject. Output verified to be valid.

It looks like, at some point, much of this will be standard across semesters, with only a few changes. At that point, some of this will move back into the base.xml file. For now, however, spring is the only one to be sorted like this, so it's only found in spring.