Standardization, minor tweaks (re-enabling some sortkeys that now mean something, or will soon ;oP and removing commented out sections of legacy code that was causing some minor file-bloat)
Fixed an error in xml-make_2007-Summer.bat (disabled the errorchecking. Summer has two data files, so will require a change to xml-compare.py, which hasn't been made yet, so it was generating an error).

commented out dtd def in xml-combine.xsl - that dtd doesn't exist, and may never exist, since it's just dsc xml and dsc's dtd is out of date. If I do fix it, I'll name it xml-combined.dtd to keep it seperate from the 'official' copy.

blah blah. I think that's it. Nothing major. I just printed out some of the code to look at, and noticed some stuff that needed to be changed. Of course, now I have to print it out again. :sigh: Poor trees.