If file issue-date.txt does not exist, and user enters valid date when prompted, the file issue-date.txt will be generated in the source/ directory.
If the file exists, but is invalid, an error will be displayed, but the file will not be overwritten (to prevent data loss). In this case, if the issue-date.txt is not part of the chatter, it can be deleted manually to allow the script to set the issue date the next time the user is prompted for and enters a valid issue date.