Reworked some of the logic in xml-fix.xsl Now both types of schedules (Summer / Non-Summer) follow the same logic path.
Added errors for classes that do not fit into any term. This may be removed in favor of creating container elements to hold this homeless data until it is accessed for output. Most of the classes falling into this group are not for display.

Fixed xml-compare.py to check Summer correctly.