Ok, minor change to code, major change to paradigm. I didn't realize, altho I should have, that the class-order in mappings take precedence over course groupings. I guess my spot checks just happened to be close enough that I didn't put two and two together. So, there's a new step: xml-split.xsl - it splits the courses based on mappings groupings. ie: if there is a course that is split around another course, there will be two courses in the data: one before containing the classes that come before, and one after that contains the classes that come after. This necessitated some shuffling in the sortkey creation, which is reflected in the updates to the dtds.
Also, I did some shell scripting to reduce the amount of change-over when the transformation/semester list changes. Now, the changes are all listed in generate-output.bat and can be changed, rerun, etc very easily. I'll work on polishing it a bit more, but it's fairly easy to run a group of transformations quickly and easily. Which is good, since nine times out of ten, the exact same group of transformations are being run every week.

I did end up creating xml-sort.xsl and it allows for significant cleanup in the actual transformations. Thus far, only output-proof.xsl has been converted, but I'll be grabbing the others as I verify output.

I'd hoped to finish this off before x-mas break, but that is looking less and less likely. I'll still push for it, but we'll see if it actually happens.