Lotsa little tweaks:
GENERAL
Found a sneaky bug - course title and long title are not used, class has a title which may or may not have anything to do with the title of the course it is in. Removed unneeded attributes from course, and threaded in the class/@title.

Adjusted the psuedo-attributes used for arranging by division/subject/etc and the is-whatever used for building the special-sections in the web output. The attributes now fall into elements within the class element. This cleans up their later removal and reduces the error-prone process of listing the attributes that should be carried over.

Finally got around to removing the .00 from @credit-hours

Added a test for difference in class/@title when restructuring the data after split & sort

WEB
Fixed a bug where a LAB w/o a LEC attached would display with only half the desired information.

Removed comments from around the ASPX preamble. While that work-around made oXygen happy, it confused IE, causing some extraordinarily weird layout effects.

Somehow, the division name got into the path for the subject output. Removed it.

PRINT
Adjusted output of COOP courses with days MTWRFSU and room TBA to display NA NA (online?)