Organized, streamlined, etc. Removed all the extra batch files (some of which never made it into the repository) and cleaned up directory structure. Might go through and rename some of the xsl files to make it more clear what's going on. Probably will add readme files to some directories for documentation purposes.
Implemented tip from Will: enrolling now grabs the date itself instead of relying on Python to provide the date for it. Removed get-date.py.

Reworked the batch files - removed option to run batch files from multiple locations. User must be in the root directory of the sandbox to run the batch files. Might add a python interface for final version, for convenience.