New utility 'cause our network sucks hard.
This is a kinda ugly little hack that just keeps the network connection open by periodically writing to a file on the dev server. It's probably going to cause fragmentation on the server, generate additional network traffic, blah blah etc blah blah. If they didn't want me to do it, they'd make the connections halfway decent to begin with.